+++ src/of-type.ts
@@ -1,10 +1,9 @@
-import { OperatorFunction } from 'rxjs'
 import { filter } from 'rxjs/operators'
 
 import { ActionCreator } from './create-action-creator'
-import { Action, AnyAction } from './create-action'
-import { getType } from './get-type'
-import { castArray } from './utils'
+import { AnyAction } from './create-action'
+import { isOfType } from './is-of-type'
+import { ExtractAction } from './types'
 
 /**
  * Filter actions emitted by the source Observable by only emitting those that
@@ -23,26 +22,8 @@ import { castArray } from './utils'
  */
 export function ofType<
   TSource extends AnyAction,
-  TActionCreator extends ActionCreator<TSource>,
->(
-  actionCreators: TActionCreator | ReadonlyArray<TActionCreator>
-): OperatorFunction<TSource, ReturnType<TActionCreator>>
-export function ofType<TSource extends AnyAction, TAction extends TSource>(
-  actions: TAction | ReadonlyArray<TAction>
-): OperatorFunction<TSource, TAction>
-export function ofType<
-  TSource extends AnyAction,
-  TType extends TSource['type']
->(
-  types: TType | ReadonlyArray<TType>
-): OperatorFunction<
-  TSource,
-  TSource extends Action<infer T> ? (T extends TType ? TSource : never) : never
->
-export function ofType(keys: ActionCreator<AnyAction> | AnyAction | string) {
-  const types = castArray(keys).map(key =>
-    typeof key === 'string' ? key : getType(key)
-  )
-
-  return filter<AnyAction>(action => types.includes(action.type))
+  TKey extends TSource['type'] | TSource | ActionCreator<TSource>,
+  TSink extends TSource = ExtractAction<TKey, TSource>
+>(keys: TKey | ReadonlyArray<TKey>) {
+  return filter<TSource, TSink>(isOfType(keys))
 }