+++ src/utils/data.ts
@@ -1,6 +1,8 @@
 export const merge = <T extends {}>(...objs: T[]): T =>
   Object.assign({}, ...objs)
 
-export function castArray<TValue>(value: TValue | TValue[]): TValue[] {
+export function castArray<TValue>(
+  value: TValue | ReadonlyArray<TValue>
+): TValue[] {
   return [].concat(<any>value)
 }